refactor(core): consolidate binary encoding
Use one Base64URL codec and SHA-256 helper.

Share it across upload tokens, content cursors, extraction identities, and document revisions.
